git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@1106 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Data Interface/Supplies.cs
@@ -40,7 +40,11 @@ namespace OpenDental{
 		///<summary>Surround with try-catch.</summary>
 		public static void DeleteObject(Supply supp){
 			//validate that not already in use.
-
+			string command="SELECT COUNT(*) FROM supplyorderitem WHERE SupplyNum="+POut.PInt(supp.SupplyNum);
+			int count=PIn.PInt(General.GetCount(command));
+			if(count>0){
+				throw new ApplicationException(Lan.g("Supplies","Supply is already in use. Not allowed to delete."));
+			}
 			DataObjectFactory<Supply>.DeleteObject(supp);
 		}
